requirements_tests.txt
flake8
pytest
pytest-asyncio
pytest-django
